# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Redoubt API
  description: |
    Redoubt is a secure, invite-only communication platform API.

    All error responses conform to RFC 9457 Problem Details for HTTP APIs.
    All timestamps are in ISO 8601 UTC format.
    All IDs are UUID v4.
  version: 0.1.0
  contact:
    name: Redoubt
  license:
    name: MIT
servers:
  - url: https://api.example.com/api/v1
    description: Production server
  - url: http://localhost:8080/api/v1
    description: Development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token
  schemas:
...
    UUID:
      type: string
      format: uuid
      example: "550e8400-e29b-41d4-a716-446655440000"
    Timestamp:
      type: string
      format: date-time
      example: "2026-02-22T10:30:00Z"

    # Error schemas (RFC 9457)
    ProblemDetails:
      type: object
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          format: uri
          description: URI reference identifying the problem type
          example: "https://redoubt.app/problems/validation-error"
        title:
          type: string
          description: Short, human-readable summary
          example: "Validation Failed"
        status:
          type: integer
          description: HTTP status code
          example: 400
        detail:
          type: string
          description: Human-readable explanation
          example: "The request body contains invalid fields"
        instance:
          type: string
          description: URI reference for this occurrence
          example: "/api/v1/auth/register"
        request_id:
          type: string
          description: X-Request-ID for tracing
          example: "abc-123-def"
        errors:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'
    FieldError:
      type: object
      required:
        - field
        - message
      properties:
        field:
          type: string
          example: "username"
        message:
          type: string
          example: "must be between 3 and 32 characters"

    # Auth schemas
...
    Space:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        name:
          type: string
          example: "My Community"
        icon_url:
          type: string
          nullable: true
        owner_id:
          $ref: '#/components/schemas/UUID'
        created_at:
          $ref: '#/components/schemas/Timestamp'
        updated_at:
          $ref: '#/components/schemas/Timestamp'
...
    Invite:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        code:
          type: string
          example: "alpha-bravo-charlie-delta"
        space_id:
          $ref: '#/components/schemas/UUID'
        created_by:
          $ref: '#/components/schemas/UUID'
        uses:
          type: integer
          example: 5
        max_uses:
          type: integer
          nullable: true
          description: null means unlimited
        expires_at:
          type: string
          format: date-time
          nullable: true
          description: null means never expires
        created_at:
          $ref: '#/components/schemas/Timestamp'
    InviteInfo:
      type: object
      properties:
        code:
          type: string
        space_name:
          type: string
        expires_at:
          type: string
          format: date-time
          nullable: true
    CreateInviteRequest:
      type: object
      properties:
        max_uses:
          type: integer
          nullable: true
          description: Maximum number of uses (null for unlimited)
        expires_in_hours:
          type: integer
          nullable: true
          description: Hours until expiry (null for never)
    InvitesListResponse:
      type: object
      properties:
        invites:
          type: array
          items:
            $ref: '#/components/schemas/Invite'

    # Health schemas
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
  responses:
    BadRequest:
      description: Bad request - validation error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Unauthorized:
      description: Unauthorized - authentication required
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Forbidden:
      description: Forbidden - insufficient permissions
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...
paths:
...
  /spaces/{id}/invites:
    get:
      tags: [Invites]
      summary: List invites
      description: List all invites for a space. Requires admin role.
      operationId: listInvites
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '200':
          description: List of invites
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvitesListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    post:
      tags: [Invites]
      summary: Create invite
      description: Create a new invite code for a space. Requires admin role.
      operationId: createInvite
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInviteRequest'
      responses:
        '201':
          description: Invite created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invite'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /invites/{code}:
    get:
      tags: [Invites]
      summary: Get invite info
      description: Get public information about an invite code. No authentication required.
      operationId: getInviteInfo
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Invite information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InviteInfo'
        '404':
          $ref: '#/components/responses/NotFound'
  /invites/{id}:
    delete:
      tags: [Invites]
      summary: Revoke invite
      description: Revoke an invite code. Requires admin role.
      operationId: revokeInvite
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '200':
          description: Invite revoked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /invites/{code}/join:
    post:
      tags: [Invites]
      summary: Join via invite
      description: Join a space using an invite code.
      operationId: joinViaInvite
      security:
        - bearerAuth: []
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Joined space
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Space'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # Health endpoints
...